requirements.txt
langchain
streamlit
python-dotenv
sentence_transformers
transformers
requests